Reworked requirements to fit new actions
Req.s 5, 7 and 8 now use `nu` instead of `mu` to combat infinite loops of starting and stopping a random conveyor to indefinitely loop.
Further, req 3 is not usable with the new synchronizations in `comm`. I want to know if we're okay with yeeting it.
Lastsly, req.s 2 and 10 are still not correct. I do not know how to write these correctly.